Add function to load audio

+++ api/src/Box9.Leds.Manager.Services/AudioPlayback/AudioTrack.cs
@@ -24,7 +24,10 @@ namespace Box9.Leds.Manager.Services.AudioPlayback
 
         public void Dispose()
         {
-            File.Delete(AudioFilePath);
+            if (File.Exists(AudioFilePath))
+            {
+                File.Delete(AudioFilePath);
+            }
         }
     }
 }
